requirements.txt
cohere>=5.0
psycopg2-binary>=2.9
pandas>=2.0
numpy>=1.24
python-dotenv>=1.0
streamlit
requests>=2.0.0
anthropic
